refactoring
Former-commit-id: 058cd19d53704a2d774754b0aaec7ec9a7194af8
Former-commit-id: 04bbbc82ed6c00fac79ab065a49e9d74b28136d0

+++ environment_model/latent_space/models_from_paper/encoder_module.py
@@ -32,5 +32,5 @@ class EncoderModule(nn.Module):
         return encoded
 
     def output_size(self):
-        return (self.output_channels,self.input_shape[1]/self.division_size, self.input_shape[2]/self.division_size)
+        return (self.output_channels,int(self.input_shape[1]/self.division_size), int(self.input_shape[2]/self.division_size))
